refactor menu directory location
