+++ Assets/Scripts/_UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e42b6f797203f654bbb1ca40afe8e24a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
